+++ src/main/java/io/muzoo/ooc/webapp/basic/servlets/RemoveUserServlet.java
@@ -1,17 +1,20 @@
 package io.muzoo.ooc.webapp.basic.servlets;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class RemoveUserServlet extends AbstractRoutableHttpServlet{
+public class RemoveUserServlet extends AbstractRoutableHttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        securityService.removeUser(request);
-        response.sendRedirect("/");
+        if(securityService.isAuthorized(request)) {
+            securityService.removeUser(request);
+            response.sendRedirect("/");
+        } else {
+            response.sendRedirect("/login");
+        }
     }
 
     @Override
